[refactor] #109 - solid 원칙에 맞도록 인터페이스 구현체로 분리하는 리팩토링 진행
