+++ src/main/java/gitbal/backend/schedule/userupdate/onedaycommit/UserOneDayCommitUpdater.java
@@ -1,0 +1,10 @@
+package gitbal.backend.schedule.userupdate.onedaycommit;
+
+import org.springframework.transaction.annotation.Transactional;
+
+public interface UserOneDayCommitUpdater {
+
+    @Transactional
+    void update();
+
+}
