+++ src/main/java/gitbal/backend/schedule/userupdate/score/UserScoreUpdater.java
@@ -1,0 +1,10 @@
+package gitbal.backend.schedule.userupdate.score;
+
+import org.springframework.transaction.annotation.Transactional;
+
+public interface UserScoreUpdater {
+
+    @Transactional
+    void update();
+
+}
